file added

+++ crud/delete.php
@@ -1,6 +1,13 @@
 <?php
 require_once "connect.php";
 
+session_start();
+if (!isset($_SESSION['login'])) {
+  header("Location: login.php");
+  exit;
+}
+
+
 $id = $_GET['id'];
 
 $sql = "DELETE FROM users WHERE id = $id";
